+++ Domitian-Engine/Force.cpp
@@ -1,7 +1,7 @@
 #include "Force.h"
 
 
-Force::Force(float myDirectionToForce,float myForceDirection, float myForce) : 
+Force::Force(double myDirectionToForce,double myForceDirection, double myForce) : 
 	force(myForce),direction_to_force(myDirectionToForce),force_direction(myForceDirection)
 {
 }
