+++ Domitian-Engine/Radian.h
@@ -6,14 +6,14 @@
 class Radian
 {
 public:
-	static float convertToGame(float myMathRadian)
+	static double convertToGame(double myMathRadian)
 	{
 		Vector2 vectored = Vector2::ToVector2(myMathRadian);
 		vectored.x = -vectored.x;
 
 		return Vector2::ToMathRadian(vectored) + 3.1459/2;
 	}
-	static float convertToMath(float myGameRadian)
+	static double convertToMath(double myGameRadian)
 	{
 		Vector2 vectored = Vector2::ToVector2(myGameRadian);
 		vectored.y = -vectored.y;
